// src/types/api.d.ts
export interface components {
  schemas: {
// ...
    SchoolCourseEntityView: {
      /** Format: int32 */
      academyYear?: number;
      categoryRefSchoolCourses?: components["schemas"]["CategoryRefEntityView"][];
      code?: string;
      /** @enum {string} */
      courseGradeType?: "A" | "AM" | "AP" | "B" | "BM" | "BP" | "C" | "CM" | "CP" | "D" | "DM" | "DP" | "F";
      /** Format: date-time */
      createdDate?: string;
      dayAndTimeRanges?: components["schemas"]["DayAndTimeRange"][];
      divisionCode?: string;
      elapsedCreatedDate?: string;
      engSubjectName?: string;
      /** @enum {string} */
      entityStatus?: "ACTIVE" | "BLOCKED" | "DELETED" | "INVALID" | "SEARCHABLE";
      extras?: string;
      /** Format: bigdecimal */
      grade?: number;
      isClosed?: boolean;
      kindsCode?: string;
      kindsName?: string;
      /** Format: int32 */
      numBooking?: number;
      planUrl?: string;
      practiceDivisionCode?: string;
      professors?: string;
      roomName?: string;
      semester?: string;
      /** Format: int32 */
      studentYear?: number;
      subDescription?: string;
      subjectName?: string;
      time?: string;
      userReports?: components["schemas"]["UserReportEntityView"][];
      uuid?: string;
    };
// ...
    TimeTableEditDto: {
      changeLog?: string;
      customCourses?: components["schemas"]["CustomCourseRegisterDto"][];
      /** @enum {string} */
      entityStatus?: "ACTIVE" | "BLOCKED" | "DELETED" | "INVALID" | "SEARCHABLE";
      isFavorite?: boolean;
      name?: string;
      /** @enum {string} */
      representativeColorType?: "BLUE" | "GREEN" | "PASTEL" | "PURPLE" | "RED" | "SPRING";
      timeTableSchoolCourses?: components["schemas"]["TimeTableSchoolCourseRegisterDto"][];
      uuid: string;
    };
    /** TimeTableEntityView */
    TimeTableEntityView: {
      /** Format: double */
      averageGrade?: number;
      colorCodes?: string[];
      /** Format: date-time */
      createdDate?: string;
      customCourses?: components["schemas"]["CustomCourseEntityView"][];
      elapsedCreatedDate?: string;
      /** @enum {string} */
      entityStatus?: "ACTIVE" | "BLOCKED" | "DELETED" | "INVALID" | "SEARCHABLE";
      isFavorite?: boolean;
      name?: string;
      /** @enum {string} */
      representativeColorType?: "BLUE" | "GREEN" | "PASTEL" | "PURPLE" | "RED" | "SPRING";
      schoolCourses?: components["schemas"]["SchoolCourseEntityView"][];
      /** @enum {string} */
      semester?: "FALL" | "SPRING" | "SUMMER" | "WINTER";
      timeTableResult?: components["schemas"]["TimeTableResult"];
      uuid?: string;
      /** Format: int32 */
      year?: number;
      yearAndSemesterName?: string;
    };
// ...
    TimeTableTemplate: {
      /** @enum {string} */
      semester?: "FALL" | "SPRING" | "SUMMER" | "WINTER";
      /** Format: int32 */
      year?: number;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}